*** empty log message ***
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@9135 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ doc/bin/makeHTML.sh
@@ -4,6 +4,7 @@
 # and implements generic generation of HTML files from BETA files.
 
 # Must be set:
+#  BETALIB  used to find source files
 #  SRC      base directory for interface files
 #  FILES    .bet files to include (releative to $SRC)
 #  MAKEINT  beta2int.perl command, including relative path
@@ -12,6 +13,12 @@
 #  INTPOST  Postprocessor applied after $MAKEINT, before $MAKEHTML
 #  OPTIONS  Extra options for $MAKEHTML 
 #  QUIET    Be quiet
+
+if [ "$BETALILB" = "" ]; then echo "$0: no BETALIB directory specified"; exit 1; fi
+if [ "$SRC"      = "" ]; then echo "$0: no SRC directory specified"; exit 1; fi
+if [ "$FILES"    = "" ]; then echo "$0: no FILES list specified"; exit 1; fi
+if [ "$MAKEINT"  = "" ]; then echo "$0: no MAKEINT script specified"; exit 1; fi
+if [ "$MAKEHTML" = "" ]; then echo "$0: no MAKEHTML script specified"; exit 1; fi
 
 /bin/rm -f *.int
